Tidy up workflow.
