+++ Cargo.toml
@@ -3,7 +3,6 @@ name = "outpack"
 version = "0.2.0"
 edition = "2021"
 rust-version = "1.70"
-include = ["schema/*"]
 
 [lib]
 crate-type = ["rlib", "cdylib"]
